requirements.txt
attrs==21.2.0
cattrs==1.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
looker-sdk==21.20.0
lxml==4.6.4
MarkupSafe==2.0.1
Pillow==8.4.0
python-docx==0.8.11
python-pptx==0.6.21
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
XlsxWriter==3.0.2
